refactor the limit and offset tests

+++ judo-runtime-core-jsl-itest/models/PagingModel/src/test/java/hu/blackbelt/judo/runtime/core/jsl/entity/PagingTest.java
@@ -307,7 +307,7 @@ public class PagingTest {
         list = itemDao
                 .query()
                 .orderBy(ItemAttribute.NUMBER)
-                .selectList(1, null);
+                .selectList(1, (Integer) null);
 
         assertEquals(1, list.size());
         assertEquals(ent2.identifier().getIdentifier(), list.get(0).identifier().getIdentifier());
@@ -315,7 +315,7 @@ public class PagingTest {
         list = itemDao
                 .query()
                 .orderBy(ItemAttribute.NUMBER)
-                .selectList(null, null);
+                .selectList(null, (Integer) null);
 
         assertEquals(2, list.size());
         assertEquals(ent2.identifier().getIdentifier(), list.get(0).identifier().getIdentifier());
@@ -324,7 +324,7 @@ public class PagingTest {
         list = itemDao
                 .query()
                 .orderBy(ItemAttribute.NUMBER)
-                .selectList(0, null);
+                .selectList(0, (Integer) null);
 
         assertEquals(2, list.size());
         assertEquals(ent2.identifier().getIdentifier(), list.get(0).identifier().getIdentifier());
@@ -333,7 +333,7 @@ public class PagingTest {
         list = itemDao
                 .query()
                 .orderBy(ItemAttribute.NUMBER)
-                .selectList(3, null);
+                .selectList(3, (Integer) null);
 
         assertEquals(2, list.size());
         assertEquals(ent2.identifier().getIdentifier(), list.get(0).identifier().getIdentifier());
